refactor(extensions): 카탈로그 스냅샷 catalog-snapshot.ts 분리 — CI 600줄 파일 가드 해소
extension-ingest-service.ts 가 734줄로 Gate 3(File Size Guard, max 600) 위반
(직전 커밋에서 이미 672줄로 위반 — 이번 판정 전면화로 표면화). 동작 변경 없는
이동: fetchCatalogSnapshot 구현 + probeInstallableAt/probeExternalEntry +
buildSkillDiscoveryPattern 을 catalog-snapshot.ts(228줄)로 추출, 서비스는
위임 메서드만 유지(544줄). buildSkillDiscoveryPattern 은 기존 import 경로
호환 재노출.
